test: 로또 도메인 테스트
