+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,0 +1,62 @@
+package lotto.domain;
+
+import lotto.util.ErrorCode;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class LottoTest {
+
+    @Test
+    void 정상입력() {
+        // given
+        List<Integer> numbers = List.of(1, 2, 3, 4, 5, 6);
+
+        // when & then
+        assertThat(new Lotto(numbers).getNumbers()).isEqualTo(numbers);
+    }
+
+    @Test
+    void 로또_번호의_개수가_6개가_넘어가면_예외가_발생한다() {
+        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage(ErrorCode.INVALID_LOTTO_NUMBER.getMessage());
+    }
+
+    @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
+    @Test
+    void 로또_번호에_중복된_숫자가_있으면_예외가_발생한다() {
+        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage(ErrorCode.DUPLICATE_LOTTO_NUMBER.getMessage());;
+    }
+
+    @Test
+    void 로또_중_범위_밖_존재() {
+        // given
+        List<Integer> numbers = List.of(1, 2, 3, 4, 5, 50);
+
+        // when & then
+        assertThatThrownBy(() -> new Lotto(numbers))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage(ErrorCode.INVALID_LOTTO_NUMBER.getMessage());
+    }
+
+    @Test
+    void toString으로_번호_변환() {
+        // given
+        List<Integer> numbers = List.of(5, 3, 6, 1, 4, 2);
+        Lotto lotto = new Lotto(numbers);
+
+        // when
+        String result = lotto.toString();
+
+        // then
+        assertThat(result).isEqualTo("[1, 2, 3, 4, 5, 6]");
+    }
+
+}
